arm64: dts: fsl-imx8mn: Enable dispmix_gpr
The DSI and eLCDIF interfaces are expecting a reference to dispmix_gpr
which is present in the i.MX8M Mini device tree, but not the i.MX8M Nano.

This patch enables the dispmix_gpr for the i.MX8M Nano which will allow
video support on the i.MX8M Nano.

Signed-off-by: Adam Ford <aford173@gmail.com>