arm64: dts: beacon-imx8mm: Fix Camera power on Beta hardware
The alpha boards had an issue where the camera power was left on during
suspend/power off, because the power was active low.

In beta hardware, the power is active high, so when the GPIO pins
go low the camera is disabled.

To help maintain backwards comatibility with older hardware, a new
device tree called beacon-imx8mm-kit-alpha which undoes the polarity
change, but the user will have to specify this device tree in the
bootloader if they have alpha boards.

Fixes: PHOENIX8-55
Signed-off-by: Adam Ford <aford173@gmail.com>