arm64: dts: beacon-imx8mm: Fix USB Host Reset
The USB controller is plumbed through a no-op USB tranceiver which
really only has a few purposes, but one of them is to be able to
reset the device.

Until now, the USB has been held out of reset using a fixed-regulator,
but this is incorrect.

This patch moves the gpio functions out of the USB's fixed regulator
into the usbphynop2 node which allows the USB system to reset the hub
as needed.  This also appears to fix an issue where the USB fails
to recover from overcurrent.

Fixes: PHOENIX8-78

Signed-off-by: Adam Ford <aford173@gmail.com>