arm64: dts: beacon-imx8mm: Fix USDHC2 CD Pin
The card detetect pin is currently setup to use the default pin,
however, it doesn't appear to be recognizing when the card is
either inserted or removed when booting from devices other than
USDHC2 (uSD card on baseboard).

This patch changes the USDCH2 CD to use GPIO, and configures the
pinmux for the for SD2_CD_B to GPIO2_IO12.

Fixes: PHOENIX8-158

Signed-off-by: Adam Ford <aford173@gmail.com>