arm64: dts: beacon-imx8mm: Enable camera reset pins
There are two GPIO pins assigned to the camera and pin muxed.
Unfortunately, only one of these two pins was being used.  The
reset pin wasn't being claimed which caused the dmesg to throw
some extra warnings.

This patch assigns the reset pin to the correct GPIO and eliminates
the message.

Signed-off-by: Adam Ford <aford173@gmail.com>