arm64: dts: beacon-imx8mm/n-lvds: Reduce LCD flicker
With the updated LVDS boards, some of the manual timing required
by Linux isn't necessary anymore because the LVDS board has a sequencer
chip that will setup the proper pins in their respective order, and
therefore it doesn't require Linux to create the artificial delays.

This patch removes the 'regulator-always-on' and regulator delays
from devices related to the LVDS screen and all helps reduce LCD
flicker during startup.

Signed-off-by: Adam Ford <aford173@gmail.com>