eeprom: Rename the LPD EEPROM decoder to beacon-productid
With Logic PD using the name of Beacon EmbeddedWorks, this patch
changes the name of the module to beacon-productid.  The word
"new" was also dropped since it's not new any more.  Since the
driver isn't unique to omap3, I removed all references to that
and replaced them as well.

This also removes the deprecated get_ds() with KERNEL_DS.

Signed-off-by: Adam Ford <aford173@gmail.com>