arm64: dts: beacon-imx8mn-spidev: Add SPIdev dts file for Nano
There is a custom device tree with spidev for the i.Mx8M Mini, but
nothing existed for Nano.  This patch introduces a comparable device
tree for the i.Mx8MN Nano beacon board which disables the EEPROM driver
to allow the SPIDEV driver to read/write to the EEPROM and it also
removes UART3, and changes the pinmux to support ECSPI1 which is
routed to header, J80.

Signed-off-by: Adam Ford <aford173@gmail.com>