arm64: dts: beacon-imx8m kits: Assign regulator to touchscreen
The touchscreen has a VDD pin, but Linux doesn't know the regulator is
needed by the touchscreen, and when there are no consumers of the
regulator, Linux powers them down.  This is a problem when the touch
controller is still expecting power, but it as gone away.

This patch adds the corresponding vdd-supply assignement to the touch
controller node to ensure the power it needs is readily available.

Fixes: PHOENIX8-214

Signed-off-by: Adam Ford <aford173@gmail.com>