arm64: dts: beacon imx8m/n kits: Correct WDOG_B pin configuration
The schematic note explicitly states to enable the internal pull-up,
but this isn't being done.

Per upstream:

"... to make sure the WDOG_B pin works properly, better to enable
internal pull up resistor
... And schmitt input is NOT necessary for this
WDOG_B output pin, so remove it; Open drain outputs provide more
flexibility to a designer as they can be pulled-up to any voltage found
in the system, so enable it as well.

Signed-off-by: Anson Huang <Anson.Huang@nxp.com>
Signed-off-by: Adam Ford <aford173@gmail.com>