arch64: dts: imx8mn: Configure spba-buses for DMA
There are several spba-buses on the SoC.  One contains
various SPI and UARTs, and the other contains a variety of
audio interfaces.  In order for the DMA to associate itself
to a Shared Peripheral Bus Arbiter (SPBA), the peripherals
need to be under a common bus associated to an spba-bus.

Signed-off-by: Adam Ford <aford173@gmail.com>