dt-bindings: imx8mn: Fix reference to IMX8MN_CLK_UART1
The entries in the dt-bindings that describe the various clocks
need to be unique.  That wasn't the case for IMX8MN_CLK_UART1.
It was using a duplicated reference which meant that attempts to
re-parent this clock were being referenced by another.
Fix this by making the clock reference unique.

Signed-off-by: Adam Ford <aford173@gmail.com>