arm64: configs: defconfig: Make i.MX SPBA subnodes modules
When adding the SPBA nodes to assign the subnodes to the SDMA
controller, it is necessary to make sure the SDMA driver and firmware
are loaded before the subnodes or it may cause instability or hanging.
Make the subnodes modules and loading the SDMA first appears to fix
the issue.

Fixes: e03dab316861 ("arm64: dts: imx8mm: Enable SPBA buses")
Signed-off-by: Adam Ford <aford173@gmail.com>