arm64: dts: imx8mm-beacon: Cleanup PCIe ref clock
The hardware does not route the negative clock request line from
the SoC to the clock source that enables the reference clock.  Instead,
the clock enable is connected to gpio expander.  Previously, we
created a 'regulator' which was really just a mechanism to enable the
pcie 100M ref clock.  Whenever the pcie regulator was turned on, it
turned on the 100MHz ref clock.  This is not ideal because in theory
the PCIe know when it needs clocks and when it doesn't so moving
this mechanism to the clock driver can clean it up.

This means creating a gpio gated clock using the 100MHz ref clock
as its source.  The new gated clock gets referenced from the
pcie controller.  When the refclock needs to be enabled, the clock
mechanism will cause the gpio to enable and disable accordingly.

This was tested by Richard F, and still yields PCIe Gen 2 performance.

Signed-off-by: Adam Ford <aford173@gmail.com>