arm64: dts: imx8mm-beacon: Fix possible PCIe refclock timeout
There needs to be a refclock for the PCIe, but if it doesn't come up
in time, it can cause a timeout and a hang.  Move the PCIe clock enable
to the regulator which has an extended delay after the power comes up
and before the PCIe enumerates to attempt to eliminate the timeout.

Fixes: PHOENIX8-355