usb: typec: hd3ss3220: Add support for controlling VBUS
Some typec devices like the ptn5110 have pins to enable sources
or sinks which can be used to enable a vbus regulator.  Unfortunately
the hd3ss3220 doesn't have such pins.  This can be emulated in
software by adding support for a vbus regulator.  Add functionality
to the hd3ss3220 to enable an optional vbus regulator when it
detects that we're in a host mode and turn it off when we disconnect
or run in device mode.