arm64: dts: imx8mp-beacon: Move USB vbus control to hd3ss3220
The phy doesn't necessarily know if the USB is running in a device
or host mode.  As such, when associated with a vbus regulator, the
regulator remains on.  This can cause an issue when the USB-C is
connected to another host and both are suppling a VBUS.

Move the VBUS regulator from the USB PHY node to the hd3ss3220
which now has the ability to automatically enable/disable the VBUS.