arm64: configs: bcn_imx8m_mn: Make camera capture drivers modules
If the Camera is loaded after the rest of the capture drivers,
the media pipeline appears to be broken.  Since the capture
drivers are built into the kernel, this always fails. To fix this,
make the capture drivers into modules.

Userspace will have to make sure the ov5640 camera
is loaded before the rest of the capture drivers