arm64: dts: imx8mp-beacon-kit: Cleanup I2C Pinmux
There are only two I2C busses that need to be enabled on the kit,
because the SOM already enables them.  Remove the duplicates, and
cleanup the pinmuxing for I2C2 and I2C4 because according to RM
bit layout, BIT3 and BIT0 are reserved.
 8  7   6   5   4   3  2 1  0
PE HYS PUE ODE FSEL X  DSE  X

Don't set the reserved bits.