arm64: dts: imx8mp-beacon: Move pinmuxing for eMMC to SOM
The SOM contains the eMMC, not the baseboard/kit.  Move the pinmux
for the eMMC to the SOM from the kit.  While this is moving,

According to RM bit layout, BIT3 and BIT0 are reserved.
 8  7   6   5   4   3  2 1  0
PE HYS PUE ODE FSEL X  DSE  X

Make sure no reseved bits are set.