arm64: dts: imx8mp-beacon-kit: Assign WM8962 CODEC to CLKOUT1
The original desire was to run the WM8962 from the SAI3 MCLK, but
the MCLK is only present when the SAI3 is running. This is a problem,
because the WM8962 needs to start the clock and calibrate which is
done before the SAI3 starts.  Since the SAI3 isn't running the MCLK
doesn't either which causes a timeout on the WM8962.
Fix this by routing the CLKOUT1 to the WM8962.  The CLKOUT1 can be
configured to run from the AUDIO_PLL2.  When the WM8962 requests
a clock, the CLKOUT1 can provide this and the CODEC can calibrate.