drm/imx: dw_hdmi-imx: Handing binding failure
In the event the dw_hdmi-imx attempts to drm_bridge_attach  before the
corresponding lcdif is available, the bind will fail.  However when
attempts to enumerate a second time cause a kernel crash, because the
original instance was never removed.  Fix this by checking the return
code of drm_bridge_attach and removing the instance if it failed,
and return an error code to defer and try again.