Input: ili210x: Enable reguator support
The datasheet used for the Ampire LCD shows the interface for the touch
controller has a VDD pin for the touch controller.

This patch adds optional regulator support for the touchscreen regulator.
This appears to resolve issues where the regulator is powered down due to
lack of consumers and keeps the regulator available when the touchscreen
controller needs it.