arm64: dts: imx8mp-beacon:  Enable FEC Interrupt
The FEC interrupts use the newer device tree style of putting the
GPIO references under the PHY node instead of the ethernet controller
node.  To enable this, move the older, deprecated style of GPIO access
for reset to the PHY node, so we can also add the IRQ as well.