drm/bridge: adv7511:  Simplify mode_valid and eliminate TODO
NXP added a TODO list theorizing that the HDMI bridge was not
able to sync certain resolutions, when in reality it was the
LCDIF that needs exact clock division or the generated signal
won't operate correctly.

Eliminate the TODO message from NXP and simiplify the mode_valid
to remove resolution options where the clock cannot divide evenly
from its parent.