arm64: dts: imx8mp-beacon: Fix PCIe Reset
When connecting to some peripherals, it appears that the reset is
not used, but other peripherals do not enumerate and/or hang.
This is because there are two PCIe connectors, but the reset-gpio
is routed to the alternate socket instead of the default socket
which appears to leave the gpio in a state holding the PCIe device
in a reset condition.

Fix this by configuring the reset-gpio to the default M.2 socket
which allows the peripherals to come out of reset and enumerate.