net: stmmac: Backport EQOS fixes from 5.15.128
The EQOS driver appears to have some timing issues when used
with some network switches where it gets an IP address and
shows the link is up, but it fails to ping the router.

Since it appears the 5.15.128 kernel has this addressed, copy
the driver from there to replace this one.

With this, testing has shown a dramatic reduction from ping
failures of 85%-100% losses happening about 68% of the times
they are attempted to <10% loss at < 10% of the time on a
router that is known to reproduce this issue.  These
same tests run an enterprise-grade switch show almost
no failures at all.