drm/bridge: adv7533: Allow IRQF_SHARED
The IRQ pin on the ADV7535 is shared with two GPIO expanders
on the i.MX8MP Beacon board, but the ADV75xx driver rejects
the IRQ request because it currently doesn't support sharing
the GPIO pin.  When it rejects the pin, the ADV75xx doesn't
enumerate and there is no video.  Enabling the shred IRQ
allows the video to work.

Signed-off-by: Adam Ford <aford@beaconembedded.com>