arm64: dts: imx8mm-beacon: Increase VPU Speed
The VPU G1, G2, and H1 are all currently set to 600MHz.

According to the datasheet, when running in overdrive mode,
the VPU_G2 can be increased to 700MHz, and the VPU_G1 and
VPU_H1 can be increased to 750MHz.  The SYS_PLL3 is already
configured for 750MHz, and it can be used as a clock source
for VPU_G1 and VPU_H1, so reset their clock parents and speed
to 750MHz.  The VPU_G2 is limited to 700MHz, so the VPU_PLL_OUT
can be used as the clock parent with its speed increased to
700MHz.

Signed-off-by: Adam Ford <aford@beaconembedded.com>