arm64: dts: imx8mp-beacon: Realign hd3ss3220 connector settings
The Mainline linux has some rules and naming convention, so pull
in the updated names and cross-reference rules.

To prevent a USB hang, move the connector from the root node and
place it as a sub-node under the hd3ss3220.  This ensures the
order of operations as the system comes out of suspend.

Signed-off-by: Adam Ford <aford@beaconembedded.com>