arm64: dts: imx8mp-beacon-kit: Make hd3ss3220 control vbus
The hd3ss3220 chip is both a USB-C mux and a Type-C detector, so it
knows if the connected device needs to run as a peripheral or if
the dev kit needs to run as a host.  Currently, the USB PHY doesn't appear
to properly handle vbus, so it appears to always be on which also means
the USB controller always shows up as a host and when connected to a host
creates a voltage contention.

Now that the hd3ss3220 driver has been updated to support enabling/disabling
vbus, update the device tree to reflect this.

Fixes: UNIFY-1293
Fixes: UNIFY-1294
Signed-off-by: Adam Ford <aford@beaconembedded.com>