arm64: dts: imx8mp: Fix Hanging on Lite versions
The pgc_vpumix power domain is controlled by the vpumix_blk_ctl,
and as such references from the pgc_vpu_g1, pgc_vpu_g2, and
pgc_vpu_vc8000e are unnecessary and they cause a hard hang on the
Lite versions because it does not exist.  Remove the references
to fix the hanging.

Signed-off-by: Adam Ford <aford@beaconembedded.com>